requirements.txt
# AI — Groq uses the OpenAI SDK (free tier at console.groq.com)
openai>=1.0.0
google-generativeai>=0.5.0
# HTTP client + scraping
httpx>=0.27.0
beautifulsoup4>=4.12.0
playwright>=1.44.0
fake-useragent>=1.5.0
# Google integrations (all free)
gspread>=6.0.0
google-auth>=2.29.0
google-api-python-client>=2.120.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
# Email verification (built on Python standard library)
pyIsEmail>=1.4.0
dnspython>=2.6.0
# Webhook server (open tracking + reply detection)
flask>=3.0.0
starkbank-ecdsa>=2.0.0
# Retry logic
tenacity>=8.3.0
# Config
python-dotenv>=1.0.0
# Database
sqlite-utils>=3.35.0
# Console output
rich>=13.7.0
# Image generation
Pillow>=10.0.0
